Fixed firestore problem. Added new functions

+++ test-app/src/components/event-list/event-list.ts
@@ -1,26 +1,27 @@
-import {Component, OnInit} from '@angular/core';
-import {EventService} from '../../services/event-service';
-import { AppEvent } from '../../interfaces/AppEvent'
-import {RouterLink} from '@angular/router';
+import { Component, inject } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterLink } from '@angular/router';
+import { EventService } from '../../services/event-service';
+import { Observable } from 'rxjs';
+import { AppEvent } from '../../interfaces/AppEvent';
 
 @Component({
   selector: 'app-event-list',
-  imports: [
-    RouterLink
-  ],
+  standalone: true,
+  imports: [CommonModule, RouterLink],
   templateUrl: './event-list.html',
-  styleUrl: './event-list.scss',
+  styleUrls: ['./event-list.scss']
 })
 export class EventList {
-  events: AppEvent[] = [];
+  private eventService = inject(EventService);
 
-  constructor(private eventService: EventService) {}
-
-  onAdd(eventToAdd: AppEvent){
-    this.eventService.addEvent(eventToAdd);
-  }
+  events$: Observable<AppEvent[]> = this.eventService.getEvents();
 
   onDelete(id: string) {
-    this.eventService.deleteEvent(id);
+    if (confirm('Ви впевнені, що хочете видалити цю подію?')) {
+      this.eventService.deleteEvent(id)
+        .then(() => console.log('Подію видалено'))
+        .catch(err => console.error('Помилка при видаленні:', err));
+    }
   }
 }
